Make action robust to updating the release PR branch (#18)
This PR makes this action robust to merging the base branch into the PR branch, after the release PR has been created. `github.event.pull_request.base.sha` changes when the base branch is merged in, so that cannot serve as our link to the commit corresponding to the workflow that created the release PR.

By using `git rev-list` and `git merge-base`, this PR allows us to always identify the `create-release-pr` workflow commit, provided that the release PR is never rebased, which we explicitly don't support.