Use the GitHub Status API instead of action success or failure (#22)
Using the success or failure of the action has drawbacks, and a status that is pending until it succeeds is more desirable. This PR implements exactly that behavior, using the GitHub Status API.

Since GitHub branch protection rules only apply to base branches, this new status will be created for every commit in every PR. Therefore, the status is set to success if the PR is not a release PR. For release PRs, the status is only set to success if the reviewer requirements are met, otherwise the status is set to pending.