Improve fetching of workflow runs (#27)
Previously, we were fetching workflow runs in a very naive manner, grabbing all workflows for all events for all branches, regardless of status. This PR ensures that we only fetch successful workflow runs triggered `workflow_dispatch` on the repository base branch. If a release branch is really old, we can still fail to fetch the workflow that created it, but that should now be exceedingly rare.