Fix action always failing for PRs from forks
Use `exit 0` instead of manually writing GitHub status to "success"
when a PR is not a release PR. This means that this action will
now be successful for all non-release PRs, including those from forks.

Explanation:
------------

GitHub actions are severely limited on PRs
(see https://github.community/t/github-actions-are-severely-limited-on-prs/18179)

This is because the GITHUB_TOKEN only ever had read-only permissions,
for security purposes, which means that the set-commit-status.sh script
cannot set the commit status to "success".

However, by default, GitHub actions will automatically
set the status to "success" anyway, if the action
finishes without any errors, so we can just change the script
to `exit 0` in this case.