Revert #37 (#42)
This reverts #37. An important claim about the implementation turned out to be false; exiting with 0 instead of explicitly setting the status check does not turn the check into a success.

It's possible to make this PR compatible with PRs from forks, but it will require using the `pull_request_target` event trigger, which adds [a slew of security concerns](https://securitylab.github.com/research/github-actions-preventing-pwn-requests/). We'll figure that out in a follow-up.